Feature: Add skeleton for class nepalinumber
