+++ nepali/number/nepalinumber.py
@@ -1,0 +1,17 @@
+"""
+Contains the class for the nepalinumber feature
+"""
+
+
+class nepalinumber:
+    """
+    Represents the nepali(devanagari) numbers and
+    the features related to arithmetic operations
+    on them
+    """
+
+    def __init__(self, value) -> None:
+        """
+        Constructor/Initializer
+        """
+        self.value = value
